解决分页部分管理的controller侧滑返回上一页的问题

+++ PersonalCenter-Extend/PersonalCenter/Center/Controllers/SegmentVC/SegmentViewController.h
@@ -12,6 +12,5 @@
 
 //子控制器的父类
 @interface SegmentViewController : UIViewController
-@property (nonatomic, assign) NSInteger topHeight;
-
+@property (nonatomic, assign) CGFloat topHeight;
 @end
